requirements.txt
arviz==0.15.1
asposestorage==1.0.2
black==23.3.0
colorama==0.4.5
cvxopt==1.3.0
Cython==0.29.33
dateparser==1.1.8
finance==0.2502
flake8==6.0.0
ipython==8.12.0
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
la==0.6.0
matplotlib==3.6.3
nbconvert==7.3.0
nbformat==5.8.0
nitime==0.10.1
numdifftools==0.9.41
numpydoc==1.5.0
packaging==21.3
pandas_datareader==0.10.0
patsy==0.5.3
prophet==1.1.1
PyGithub==1.58.1
Pygments==2.13.0
pymc==5.2.0
pymc3==3.11.5
python_dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
rpy==0.99.2
rpy2==3.5.10
scikit_learn==1.2.2
scikits.talkbox==0.2.5
seaborn==0.12.2
setuptools==65.5.0
Sphinx==6.1.3
sphinx_material==0.0.35
sphinxcontrib_programoutput==0.17
streamlit==1.16.0
sympy==1.11.1
tabular==0.1
theano==1.0.5
traitlets==5.9.0
typing_extensions==4.5.0
virtualenv==20.21.0
xgboost==1.7.3
yapf==0.32.0
